pyproject.toml
[project]
name = "skills"
version = "0.1.0"
description = "Plugin-first skills workspace for Claude Code and other skill-compatible agents."
requires-python = ">=3.12"
dependencies = [
  "caldav",
  "google-api-python-client",
  "google-auth-httplib2",
  "google-auth-oauthlib",
  "icalendar",
]

[project.optional-dependencies]
agentmail = [
  "agentmail",
]

[tool.uv]
package = false
